+++ resources/views/membro/regiao.blade.php
@@ -3,8 +3,8 @@
 @section('nivel2', '<li class="active">Regiões</li>')
 
 @section('content')
+    @can('membro-regiao-create')
     {!! Form::open(array('route'=>'regiao.store','class'=>'form-inline')) !!}
-
     <div class="form-group{{ $errors->has('nome') ? ' has-error' : '' }}">
         {!! Form::label('nome', 'Nome da região:' , ['class'=>'control-label'])!!}
 
@@ -15,11 +15,12 @@
             </span>
         @endif
 
-        {!! Form::submit('Adicionar',['class'=>'btn btn-primary']) !!}
 
     </div>
+    {!! Form::submit('Adicionar',['class'=>'btn btn-primary']) !!}
 
     {!! Form::close() !!}
+    @endcan
 
     <hr class="divider">
 
@@ -37,7 +38,7 @@
                             <tr>
                                 <th>#</th>
                                 <th>Nome</th>
-                                <th>Remover</th>
+                                @can('membro-regiao-remove')<th>Remover</th>@endcan
                             </tr>
                         </thead>
                         <tbody>
@@ -46,13 +47,13 @@
                             <tr class="grupo-ativo">
                                 <th scope="row" title="{{ $regiao->id }}">{{$i++}}</th>
                                 <td class="nome-grupo">{{ $regiao->nome }}</td>
-                                <td>
+                                @can('membro-regiao-remove')<td>
                                     {{ Form::open([ 'method'  => 'delete',
                                         'route' => [ 'regiao.destroy', $regiao->id ] ]) }}
                                         {{ Form::button('<i class="fa fa-trash-o" aria-hidden="true"></i>',
                                             ['class' => 'btn btn-danger','type'=>'submit']) }}
                                             {{ Form::close() }}
-                                        </td>
+                                        </td>@endcan
                                     </tr>
                                 @endforeach
                             </tbody>
